refactor: Icon 컴포넌트 폴더링
