--- src/components/atoms/IconEyeClosed/index.js
@@ -1,34 +1,0 @@
-import PropTypes from 'prop-types';
-import { COLOR } from '../../../constants';
-
-export const IconEyeClosed = (props) => {
-  const { width, color, ...rest } = props;
-
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width={width} viewBox="0 0 16 16" fill="none" {...rest}>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M10.6639 7.87579C10.6009 6.50187 9.49818 5.39918 8.12425 5.33618L10.6639 7.87579ZM5.53249 6.98706C5.40415 7.29937 5.33337 7.64143 5.33337 8.00001C5.33337 9.47277 6.52728 10.6667 8.00004 10.6667C8.35862 10.6667 8.70067 10.5959 9.01299 10.4676L8.20021 9.65478C8.13458 9.66263 8.06778 9.66667 8.00004 9.66667C7.07957 9.66667 6.33337 8.92048 6.33337 8.00001C6.33337 7.93226 6.33742 7.86547 6.34527 7.79984L5.53249 6.98706Z"
-        fill={color}
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M12.8757 10.0877C13.1261 9.84152 13.3497 9.60165 13.543 9.38236L13.5995 9.31836C13.9414 8.93222 14.2618 8.57015 14.2618 8C14.2618 7.42985 13.9414 7.06778 13.5995 6.68164L13.543 6.61764C13.0042 6.00625 12.2296 5.23485 11.294 4.61323C10.3605 3.99305 9.23336 3.5 8 3.5C7.48565 3.5 6.98976 3.58575 6.5187 3.73071L7.34576 4.55778C7.56169 4.52012 7.78001 4.5 8 4.5C8.96664 4.5 9.90123 4.8885 10.7406 5.44615C11.5778 6.00236 12.2871 6.70507 12.7927 7.2788C13.2226 7.76657 13.2618 7.84601 13.2618 8C13.2618 8.15399 13.2226 8.23343 12.7927 8.7212C12.6101 8.92848 12.4008 9.15259 12.1685 9.38055L12.8757 10.0877ZM4.53247 5.98713C4.0092 6.41828 3.55895 6.87974 3.20727 7.2788C2.77742 7.76657 2.73816 7.84601 2.73816 8C2.73816 8.15399 2.77742 8.23343 3.20727 8.7212C3.71288 9.29493 4.42223 9.99764 5.25942 10.5538C6.09878 11.1115 7.03336 11.5 8 11.5C8.57536 11.5 9.13936 11.3624 9.67869 11.1333L10.4321 11.8868C9.69562 12.2521 8.87451 12.5 8 12.5C6.76664 12.5 5.63951 12.007 4.70604 11.3868C3.7704 10.7652 2.99582 9.99375 2.45703 9.38236L2.40046 9.31836C2.05865 8.93222 1.73816 8.57015 1.73816 8C1.73816 7.42985 2.05865 7.06778 2.40046 6.68164L2.45703 6.61764C2.81525 6.21116 3.27768 5.73395 3.82239 5.27704L4.53247 5.98713Z"
-        fill={color}
-      />
-      <path d="M3.33337 1.33331L14 12" stroke={color} />
-    </svg>
-  );
-};
-
-IconEyeClosed.propTypes = {
-  width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
-  color: PropTypes.string.isRequired,
-};
-
-IconEyeClosed.defaultProps = {
-  width: 16,
-  color: COLOR.ICON_DEFAULT,
-};
