+++ src/components/atoms/Icons/IconLogo/index.js
@@ -1,0 +1,25 @@
+import PropTypes from 'prop-types';
+import { COLOR, LAYOUT } from '../../../../constants';
+
+export const IconLogo = (props) => {
+  const { width, color, ...rest } = props;
+
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" width={width} viewBox="0 0 24 24" fill="none" {...rest}>
+      <path
+        d="M12 2C7.58 2 4 2.5 4 6V15.5C4 17.43 5.57 19 7.5 19L6 20.5V21H18V20.5L16.5 19C18.43 19 20 17.43 20 15.5V6C20 2.5 16.42 2 12 2ZM7.5 17C6.67 17 6 16.33 6 15.5C6 14.67 6.67 14 7.5 14C8.33 14 9 14.67 9 15.5C9 16.33 8.33 17 7.5 17ZM11 11H6V6H11V11ZM16.5 17C15.67 17 15 16.33 15 15.5C15 14.67 15.67 14 16.5 14C17.33 14 18 14.67 18 15.5C18 16.33 17.33 17 16.5 17ZM18 11H13V6H18V11Z"
+        fill={color}
+      />
+    </svg>
+  );
+};
+
+IconLogo.propTypes = {
+  width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
+  color: PropTypes.string.isRequired,
+};
+
+IconLogo.defaultProps = {
+  width: LAYOUT.NAVBAR.LOGO_WIDTH,
+  color: COLOR.TEXT.NAVBAR,
+};
