+++ src/components/atoms/Icons/IconSetting/index.js
@@ -1,0 +1,25 @@
+import PropTypes from 'prop-types';
+import { COLOR, LAYOUT } from '../../../../constants';
+
+export const IconSetting = (props) => {
+  const { width, color, ...rest } = props;
+
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" width={width} viewBox="0 0 12 12" fill="none" {...rest}>
+      <path
+        d="M9.57001 6.46995C9.59001 6.31995 9.60001 6.16495 9.60001 5.99995C9.60001 5.83995 9.59001 5.67995 9.56501 5.52995L10.58 4.73995C10.67 4.66995 10.695 4.53495 10.64 4.43495L9.68001 2.77495C9.62001 2.66495 9.49501 2.62995 9.38501 2.66495L8.19001 3.14495C7.94001 2.95495 7.67501 2.79495 7.38001 2.67495L7.20001 1.40495C7.18001 1.28495 7.08001 1.19995 6.96001 1.19995H5.04001C4.92001 1.19995 4.82501 1.28495 4.80501 1.40495L4.62501 2.67495C4.33001 2.79495 4.06001 2.95995 3.81501 3.14495L2.62001 2.66495C2.51001 2.62495 2.38501 2.66495 2.32501 2.77495L1.37001 4.43495C1.31001 4.53995 1.33001 4.66995 1.43001 4.73995L2.44501 5.52995C2.42001 5.67995 2.40001 5.84495 2.40001 5.99995C2.40001 6.15495 2.41001 6.31995 2.43501 6.46995L1.42001 7.25995C1.33001 7.32995 1.30501 7.46495 1.36001 7.56495L2.32001 9.22495C2.38001 9.33495 2.50501 9.36995 2.61501 9.33495L3.81001 8.85495C4.06001 9.04495 4.32501 9.20495 4.62001 9.32495L4.80001 10.595C4.82501 10.715 4.92001 10.8 5.04001 10.8H6.96001C7.08001 10.8 7.18001 10.715 7.19501 10.595L7.37501 9.32495C7.67001 9.20495 7.94001 9.04495 8.18501 8.85495L9.38001 9.33495C9.49001 9.37495 9.61501 9.33495 9.67501 9.22495L10.635 7.56495C10.695 7.45495 10.67 7.32995 10.575 7.25995L9.57001 6.46995ZM6.00001 7.79995C5.01001 7.79995 4.20001 6.98995 4.20001 5.99995C4.20001 5.00995 5.01001 4.19995 6.00001 4.19995C6.99001 4.19995 7.80001 5.00995 7.80001 5.99995C7.80001 6.98995 6.99001 7.79995 6.00001 7.79995Z"
+        fill={color}
+      />
+    </svg>
+  );
+};
+
+IconSetting.propTypes = {
+  width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
+  color: PropTypes.string.isRequired,
+};
+
+IconSetting.defaultProps = {
+  width: LAYOUT.NAVBAR.ITEM_WIDTH,
+  color: COLOR.TEXT.NAVBAR,
+};
