+++ src/components/index.js
@@ -3,22 +3,22 @@ export { ButtonRadio } from './atoms/ButtonRadio';
 export { ButtonRound } from './atoms/ButtonRound';
 export { ButtonSquare } from './atoms/ButtonSquare';
 export { ColorPicker } from './atoms/ColorPicker';
-export { IconArrowLTR } from './atoms/IconArrowLTR';
-export { IconCheck } from './atoms/IconCheck';
-export { IconDropDown } from './atoms/IconDropDown';
-export { IconEyeClosed } from './atoms/IconEyeClosed';
-export { IconEyeOpen } from './atoms/IconEyeOpen';
-export { IconLock } from './atoms/IconLock';
-export { IconLogo } from './atoms/IconLogo';
-export { IconMail } from './atoms/IconMail';
-export { IconPath } from './atoms/IconPath';
-export { IconPerson } from './atoms/IconPerson';
-export { IconPlus } from './atoms/IconPlus';
-export { IconSearch } from './atoms/IconSearch';
-export { IconSetting } from './atoms/IconSetting';
-export { IconSubway } from './atoms/IconSubway';
-export { IconTrashCan } from './atoms/IconTrashCan';
-export { IconWindow } from './atoms/IconWindow';
+export { IconArrowLTR } from './atoms/Icons/IconArrowLTR';
+export { IconCheck } from './atoms/Icons/IconCheck';
+export { IconDropDown } from './atoms/Icons/IconDropDown';
+export { IconEyeClosed } from './atoms/Icons/IconEyeClosed';
+export { IconEyeOpen } from './atoms/Icons/IconEyeOpen';
+export { IconLock } from './atoms/Icons/IconLock';
+export { IconLogo } from './atoms/Icons/IconLogo';
+export { IconMail } from './atoms/Icons/IconMail';
+export { IconPath } from './atoms/Icons/IconPath';
+export { IconPerson } from './atoms/Icons/IconPerson';
+export { IconPlus } from './atoms/Icons/IconPlus';
+export { IconSearch } from './atoms/Icons/IconSearch';
+export { IconSetting } from './atoms/Icons/IconSetting';
+export { IconSubway } from './atoms/Icons/IconSubway';
+export { IconTrashCan } from './atoms/Icons/IconTrashCan';
+export { IconWindow } from './atoms/Icons/IconWindow';
 export { Input } from './atoms/Input';
 export { Modal } from './atoms/Modal';
 export { Section } from './atoms/Section';
